web-preferences-api/reducedTransparency.html

<!DOCTYPE html>
<html>
  <head>
    <meta charset="utf-8" />
    <title>WebPreferences Test: reducedTransparency attribute tests</title>
    <link rel="author" title="Luke Warlow" href="mailto:luke@warlow.dev" />
    <link rel="help" href="https://wicg.github.io/web-preferences-api/" />
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
  </head>
  <body>
    <script>
      test(() => {
        assert_true('preferences' in navigator, 'navigator.preferences defined');
        assert_true('reducedTransparency' in navigator.preferences, 'navigator.preferences.reducedTransparency defined');
      }, "navigator.preferences.reducedTransparency supported");
      test(() => {
        navigator.preferences.reducedTransparency = 'no-preference';
        assert_true(window.matchMedia('(prefers-reduced-transparency: no-preference)').matches, '(prefers-reduced-transparency: no-preference) matches');
        navigator.preferences.reducedTransparency = 'reduce';
        assert_true(window.matchMedia('(prefers-reduced-transparency: reduce)').matches, '(prefers-reduced-transparency: reduce) matches');
        navigator.preferences.reducedTransparency = null;
        assert_true(window.matchMedia('(prefers-reduced-transparency: no-preference)').matches, '(prefers-reduced-transparency: no-preference) matches');
      }, "navigator.preferences.reducedTransparency updates (prefers-reduced-transparency)");
      test(() => {
        assert_throws_js(TypeError, () => { navigator.preferences.reducedTransparency = 'this-is-an-invalid-value'; });
        assert_equals(null, navigator.preferences.reducedTransparency);
        navigator.preferences.reducedTransparency = null;
      }, "navigator.preferences.reducedTransparency set to invalid value throws TypeError");
    </script>
  </body>
</html>
